requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
psutil==6.0.0
python-dotenv==1.0.1
aiofiles>=23.2.1         
anyio>=4.2.0           
pydantic>=2.8.0          
starlette>=0.38.2,<0.39  
orjson>=3.9.15         
